fix(hub): evict the interpreter-reaching half of graph_rendering, and declare the four packages that were clean but unpoliced
`graph/graph_rendering.py` reached `interpreter_hub` through
`pipeline.dry_run_pipeline` — 77 interpreter modules — with every gate
green. The reason is the finding: both the layer rule and the transitive
rule filter their candidates through `is_runtime_layer`, and
`pipelex.graph` was not declared, so the module sat outside the rules'
domain entirely. An undeclared package is not neutral, it is unpoliced.

Split along the boundary's own question, "do I need a loaded method?":
the pure renderer (`GraphFormat`, `render_graph_from_spec`) stays in
`pipelex.graph`; the bundle-driven half that must dry-run a method first
(`generate_graph_for_bundle`, `generate_view_for_bundle`) becomes
`pipelex.pipeline.bundle_graph_rendering`, next to the `dry_run_pipeline`
it wraps.

`pipelex.graph`, `pipelex.tracing`, `pipelex.observer` and
`pipelex.errors` join `RUNTIME_LAYER_PACKAGES` — each measured to load
zero interpreter modules — which is what puts the evicted module's
successor in the guard's domain. Negative control: re-adding the import
now fails the guard with `interpreter-hub-transitive`; before the
declaration it produced no finding at all. A membership test pins the
four names, because deleting one otherwise leaves the suite green.

Test placement follows: the four bundle-half modules move to
`tests/unit/pipelex/pipeline/`, `test_graph_rendering.py` moves from
`tests/unit/pipelex/cli/` to `tests/unit/pipelex/graph/`, and the
targeted-test map gains the unit path for `pipelex/pipeline/`.
