+++ .drift/acks/hub-layering-convention.toml
@@ -1,10 +1,10 @@
 contract = "hub-layering-convention"
-digest = "sha256:19d8b6737dc50fdc11f754b6077ae0b41d0c3365a85b3407ab4326708b8efab4"
-reviewed_by = "Louis Choquel"
-reviewed_at = "2026-07-27T17:01:19Z"
-rationale = "M2: pipelex/plugins/ split into the plugin mechanism (stays) and the seventeen built-in vendor adapters (now pipelex/providers/). Trigger diff on hub_layering_guard.py is the new 'pipelex.providers' entry plus a rewritten note above the tuple. Reviewed the target for real and edited five places. (1) The layer-rule enumeration under Enforcement listed seven declared packages; it now lists eight. This is the doc-enumeration rot M1 flagged, and it was in a section the diff did not otherwise touch. (2) 'Where the built-in plugins split' claimed pipelex/plugins/ is where the runtime-port adapters live — the exact sentence M2 falsifies. Rewritten to name the two splits separately, because they are routinely conflated: mechanism-vs-adapters does NOT cross a layer (both packages are runtime-layer, the split makes the one-way dependency legible), while runtime-vs-interpreter does. Added the sentence third-party plugin authors need: pipelex.plugins.contract and pipelex.plugins.registrar did not move, and the 'pipelex.plugins' entry-point group is a group identifier, not a module path. (3) The import-closure section named plugins.builtins as the entry point and justified it by history; the module is now providers.builtins, so the paragraph was rewritten to keep the history true under the old name and to state why one entry point still covers both packages. (4) Known inversions gained two bullets recording the cogt -> vendor edges as deliberate: the four config imports are accepted permanently (making vendor config sections plugin-contributed would trade the statically-typed configs.py/pipelex.toml sync for a dynamic registry), while the two img-gen factory imports are slated to move inward to a neutral cogt/img_gen/ mapping module — no registry, because dispatch is keyed on the cogt-owned AspectRatioTaxonomy and the gateway worker is one SDK spanning both taxonomies, so no provider key exists to dispatch on. (5) The fixed worked example still pointed at plugins/builtins.py; annotated with its current path. Verified the remaining claims rather than assuming: 'plugins -> runtime_bridge' still names only mechanism modules (bundle_validator_registry, orchestrator_registry, registrar — all three stayed), and the 'Why the boundary exists' must-not-name list plus the verification snippet enumerate the INTERPRETER packages, which M2 does not touch, so both are correct unchanged. Ran the three prescribed mechanical checks by script: all 8 declared packages named under Enforcement, all 27 runtime_hub and 22 interpreter_hub get_* accessors covered by the partition tables, and the matched triple (guard tuple / closure INTERPRETER_PACKAGES / doc INTERPRETER snippet) agrees. As in every prior entry the mechanical checks passed while blind to all five findings above. One defect they were also blind to, found only by reading the generated output: _SUBSYSTEM_SECTIONS in error_pages_generator.py keys on the module's SECOND segment, so every vendor error class silently reclassified from 'Inference & providers' to the fallback 'Platform & tooling' macro — the identical shape to M1's mthds_parsing regression, three weeks later, caught this time because M1's lesson was to read the gep diff rather than its exit code."
+digest = "sha256:541b794679c211f259ca44cc0c22762ec5d0a62e4841abfcc852e89575c37a67"
+reviewed_by = "Claude (Opus 5)"
+reviewed_at = "2026-07-29T11:53:55Z"
+rationale = "Trigger is hub_layering_guard.py: RUNTIME_LAYER_PACKAGES gains pipelex.graph, pipelex.tracing, pipelex.observer and pipelex.errors, and the note above the tuple is extended to say why each is there. Reviewed the target for real and edited two places. (1) The layer-rule enumeration under Enforcement listed eight declared packages and now lists twelve — the same doc-enumeration rot M1 and M2 each hit, in a section the trigger diff does not otherwise touch — plus a new paragraph naming what each of the four packages is and why kit and language stay out. (2) Known inversions gained the fixed case this change found: graph/graph_rendering.py reached interpreter_hub through pipeline.dry_run_pipeline with every gate green, because rules 1 and 3 both filter candidates through is_runtime_layer and pipelex.graph was undeclared, so the module was outside the rules' domain entirely. It is the pipelex.plugins worked example one level up — not an overreaching declaration but no declaration at all — and it is recorded next to it with the generalized lesson this track produced: an undeclared package is not neutral, it is unpoliced. The remedy was placement, per the same playbook: the bundle-driven half is now pipelex.pipeline.bundle_graph_rendering, the pure renderer stayed. Ran the two prescribed mechanical checks by script: all twelve declared packages named under The rule, and all 32 runtime_hub / 31 interpreter_hub public accessors covered by the partition tables. Both passed while blind to finding (1), as in every prior entry. Negative control run and recorded in the plan: re-adding the evicted import makes the guard fail with interpreter-hub-transitive; before the declaration it produced no finding at all. Verified the rest rather than assuming: the 'Why the boundary exists' must-not-name list and the verification snippet enumerate the INTERPRETER packages, which this change does not touch, and the cogt-neutrality bullet's claim that both sides are runtime-layer is unaffected."
 
 [trigger_files]
-"pipelex/cli/dev_cli/commands/hub_layering_guard.py" = "blob:47fef9ff0a052216b7925c6d372226aa47319fa2"
+"pipelex/cli/dev_cli/commands/hub_layering_guard.py" = "blob:819d32842b4303ef0ed78170bc23288e39d06068"
 "pipelex/interpreter_hub.py" = "blob:b7ac885d348d32f1aa4b1e660da5b03533fb305b"
 "pipelex/runtime_hub.py" = "blob:33fc3099cc4ab0c741fefb899bb93ee6b91c7cf7"
