+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Changed
 
+- **Breaking: the bundle-driven half of `graph_rendering` moved to `pipelex.pipeline`, and four measured-clean packages joined the declared runtime layer.** `pipelex.graph.graph_rendering.generate_graph_for_bundle` / `generate_view_for_bundle` are now `pipelex.pipeline.bundle_graph_rendering.*`; `GraphFormat` and `render_graph_from_spec` stay put. The split is along "do I need a loaded method?", which is the hub boundary's own question — producing a graph *spec* from a bundle requires dry-running it, rendering an existing spec does not. It matters because the old module reached `interpreter_hub` (and ~76 interpreter modules) through `pipeline.dry_run_pipeline` while every gate stayed green: the layer rule and the transitive rule both filter their candidates through `is_runtime_layer`, and `pipelex.graph` was undeclared, so the module sat outside the rules' domain entirely. `pipelex.graph`, `pipelex.tracing`, `pipelex.observer` and `pipelex.errors` are now in `RUNTIME_LAYER_PACKAGES` — each measured to load zero interpreter modules — which is what puts the evicted module's successor in the guard's domain. The generalized lesson, now recorded in [Known inversions](contribute/hub-layering.md#known-inversions): **an undeclared package is not neutral, it is unpoliced.** `pipelex.kit` (data files) and `pipelex.language` (has interpreter edges) stay out.
 - **Breaking: `pipelex.hub` is gone, split into `pipelex.runtime_hub` and `pipelex.interpreter_hub`.** The split names the two layers it separates, on the textbook language-implementation division: the **runtime** is the machinery present at execution time whatever is loaded, the **interpreter** reads a method and executes it. There is no alias shim — a stale import fails loudly at import time rather than silently resolving to the wrong layer. `runtime_hub` holds process-scoped infrastructure (config, console, secrets, storage, telemetry, models, inference, content generation, reporting, run mode, tracing, the plugin registries, `get_class_registry`); `interpreter_hub` holds the loaded method's machinery (library manager and lookups, the current-library contextvar, library dirs, pipe router, pipe run and pipeline manager, the PipeFunc executor). `get_pipelex_hub` splits into `get_runtime_hub` / `get_interpreter_hub`, so every external call site must pick the container it actually meant. The rule is one-way and mechanically enforced (`make check-hub-layering`, alias `chl`, in `make agent-check` and CI): `interpreter_hub` may import `runtime_hub`, never the reverse, and no module anywhere may reference `pipelex.hub` — in an import *or* a string literal (dynamic import, mock patch target, config path). The payoff is measurable and pinned by a subprocess import-closure test: importing the Pipelex runtime now loads **zero** interpreter modules, down from 50 — you can embed the inference engine without loading a line of the MTHDS interpreter. See [Hub layering convention](contribute/hub-layering.md).
 - **Breaking: misplaced types moved to the package that actually owns them.** `JobMetadata` / `JobCategory` / `UnitJobId` → `pipelex.system.job_metadata`, `JobMetadataError` → `pipelex.system.exceptions`, `TraceContext` → `pipelex.system.trace_context`, `DataInclusionConfig` → `pipelex.system.data_inclusion_config`, `TextFormat` → `pipelex.tools.templating.text_format`, `TemplatingStyle` / `TagStyle` → `pipelex.tools.templating.templating_style`, `TemplateCategory` → `pipelex.tools.jinja2.template_category`, and the neutral resolved-field layer → `pipelex.core.concepts.resolved_fields`. The TOML config shape is unchanged throughout (`[...graph_config.data_inclusion]` still parses); only the Python homes moved. The two pipe renderers regrouped into one `rendering/` package (`input_renderer` from `core.pipes.inputs`, `output_renderer` from `core.pipes.output`); they ship at `pipelex.pipe_machinery.rendering`, having left `core/` with the rest of the Pipe machinery — see the entry below.
 - **Breaking: `core/`'s data model resolves concepts and pipes through injected providers instead of reaching into the hub.** `StuffFactory.make_stuff_from_stuff_content_or_data` / `make_from_blueprint` / `make_from_concept_ref`, `InputShaper.shape` / `resolve_input_kind`, `WorkingMemoryFactory.make_from_pipeline_inputs`, `InputStuffSpecsFactory.make_from_blueprint` / `make_from_string`, and `StuffSpecFactory.make_from_blueprint` all take a required `concept_provider`. A new read-side contract owns that surface — `ConceptProviderAbstract` (`pipelex.core.concepts.concept_provider_abstract`), which `ConceptLibraryAbstract` now extends, keeping library management in the interpreter layer. There is no pipe counterpart: nothing in core's runtime half resolves a pipe, so `get_required_pipe` stays on `PipeLibraryAbstract`. `pipelex.core`'s data-model packages join the enforced runtime layer as a result; the half that names a `Pipe` stays in the interpreter layer by construction.
